pyproject.toml
[project]
name = "trainer"
version = "0.1.0"
description = "An open-source agentic AI personal trainer for fitness and health"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.2.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
trainer = "trainer.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "integration: integration tests requiring external services",
    "unit: unit tests that don't require external services",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
